added --extra option to fuel-download and fuel-convert
Added the ability to indicate an extra downloader package
by name from the command line of fuel-download and
fuel-convert. To enforce the constraints of subparsers
that cannot be known until after extras is processed,
a simple argv pre-processor was added for "--extra".
This option is also still added to the argparse parser
so that help messages appear as expected.

Note that these same extra packages can still be
set more permanently by adding them to
fuel.config.extra_downloaders and fuel.config.extra_converters

Expected usage:

    pip install my_pkg
    fuel-download --extra my_pkg mydata [--mydata_option foo]
    fuel-convert  --extra my_pkg mydata [--mydata_option foo]